build: finish initial build of post route

+++ db/postQueries.js
@@ -1,11 +1,12 @@
 const { PrismaClient } = require("@prisma/client");
+const { comment } = require("./queries");
 
 const prisma = new PrismaClient();
 
 async function create(title, authorId, contentString, publishedBoolean) {
   const content = contentString.length > 0 ? contentString : null;
   const published = publishedBoolean ? publishedBoolean : false;
-  await prisma.post.create({
+  const post = await prisma.post.create({
     data: {
       title,
       content,
@@ -13,16 +14,56 @@ async function create(title, authorId, contentString, publishedBoolean) {
       published,
     },
   });
+
+  let success;
+
+  if (!post) {
+    success = false;
+  } else {
+    success = true;
+  }
+
+  return success;
+}
+
+async function readSingle(id) {
+  const post = await prisma.post.findFirst({
+    where: {
+      id,
+    },
+  });
+
+  if (!post) {
+    return null;
+  }
+
+  return post;
 }
 
 async function deleteSingle(id) {
-  await prisma.post.delete({
-    where: { id },
-  });
+  const post = await readSingle(id);
+
+  if (post) {
+    const postComments = await prisma.comment.findMany({
+      where: { postId: id },
+    });
+    for (comment in postComments) {
+      const commentId = comment.id;
+      await prisma.comment.delete({
+        where: { id: commentId },
+      });
+    }
+
+    await prisma.post.delete({
+      where: { id },
+    });
+    return true;
+  }
+  return false;
 }
 
 async function readRecent(qty) {
-  qty = qty > 0 ? qty : 10;
+  qty = qty ? qty : 10;
   const posts = await prisma.post.findMany({
     orderBy: {
       createdAt: "desc",
@@ -33,16 +74,6 @@ async function readRecent(qty) {
   return posts;
 }
 
-async function readSingle(id) {
-  const post = await prisma.post.findFirst({
-    where: {
-      id,
-    },
-  });
-
-  return post;
-}
-
 async function update(id, newTitle, newContent, newPublished) {
   const oldPost = await readSingle(id);
   const title = newTitle.length > 0 ? newTitle : oldPost.title;
@@ -50,7 +81,7 @@ async function update(id, newTitle, newContent, newPublished) {
   const published =
     newPublished !== undefined ? newPublished : oldPost.published;
 
-  await prisma.post.update({
+  const updatedPost = await prisma.post.update({
     where: { id },
     data: {
       title,
@@ -58,6 +89,7 @@ async function update(id, newTitle, newContent, newPublished) {
       published,
     },
   });
+  return updatedPost;
 }
 
 module.exports = {
